+++ modules/transformer/src/babel/index.ts
@@ -3,6 +3,7 @@ import * as parser from '@babel/parser';
 import * as traverse from '@babel/traverse';
 import * as presets from './presets';
 import * as plugins from './plugins';
+import * as helpers from './helper';
 
 export {
     core,
@@ -10,4 +11,5 @@ export {
     traverse,
     presets,
     plugins,
+    helpers,
 };