+++ test/test-engine-source/type-merge/type.web.ts
@@ -1,0 +1,4 @@
+export class A {
+    testProperty: number = 1;
+    test (): void {}
+}